* removed parameter plot_aspect_ratio since its a duplicate of plotaspectratio * rename Figure to FigureEnvelope * added testcase for github 37
