+++ tests/asserts/asserts.py
@@ -5,4 +5,4 @@ def assert_num_tabs(figs, *args):
 
 def assert_num_figures(figs, *args):
     for idx, num in enumerate(args):
-        assert len(figs[idx][0].figures) == num
+        assert len(figs[idx][0].figure_envs) == num
